Moved code to services/contextnavigation making it more production ready
