+++ eea/volto/policy/restapi/configure.zcml
@@ -1,6 +1,7 @@
 <configure xmlns="http://namespaces.zope.org/zope">
   <include package="plone.restapi" />
   <include package=".deserializer" />
-  <include package=".serializer" />
   <include package=".navigation" />
+  <include package=".serializer" />
+  <include package=".services" />
 </configure>
