+++ eea/volto/policy/restapi/services/__init__.py
@@ -1,0 +1,1 @@
+"""RestAPI services"""
