+++ eea/volto/policy/restapi/services/configure.zcml
@@ -1,0 +1,6 @@
+<configure
+    xmlns="http://namespaces.zope.org/zope"
+    xmlns:zcml="http://namespaces.zope.org/zcml"
+    >
+  <include package=".contextnavigation" />
+</configure>
