+++ eea/volto/policy/restapi/services/contextnavigation/__init__.py
@@ -1,0 +1,1 @@
+""" RestAPI context navigation services """
