+++ eea/volto/policy/restapi/services/contextnavigation/configure.zcml
@@ -1,0 +1,20 @@
+<configure
+    xmlns="http://namespaces.zope.org/zope"
+    xmlns:plone="http://namespaces.plone.org/plone"
+    >
+
+  <adapter
+      factory=".get.EEAContextNavigation"
+      name="contextnavigation"
+      />
+
+  <plone:service
+      method="GET"
+      factory=".get.EEAContextNavigationGet"
+      for="zope.interface.Interface"
+      permission="zope2.View"
+      name="@contextnavigation"
+      layer="plone.restapi.interfaces.IPloneRestapiLayer"
+      />
+
+</configure>
